@@ -5,18 +5,22 @@
     "license": "MIT",
     "homepage": "https://www.vrok.de/",
     "keywords": ["latex", "pdflatex", "wrapper"],
-    "minimum-stability": "dev",
 
     "require": {
-        "php": ">=7.1"
+        "php": "^7.4"
     },
     "require-dev": {
-        "phpunit/phpunit": "^6.3.1",
-        "zendframework/zend-coding-standard": "~1.0.0"
+        "friendsofphp/php-cs-fixer":  "^2.16.3",
+        "phpunit/phpunit": "^9.2.3"
     },
     "autoload": {
         "psr-4": {
             "TexWrapper\\": "src/"
+        }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "TexWrapper\\Tests\\": "tests/"
         }
     },
     "scripts": {
@@ -24,10 +28,9 @@
             "@cs-check",
             "@test"
         ],
-        "cs-check": "phpcs",
-        "cs-fix": "phpcbf",
+        "cs-check": "php-cs-fixer fix --dry-run --diff -v",
+        "cs-fix": "php-cs-fixer fix -v",
         "test": "phpunit --colors=always",
-        "test-coverage": "phpunit --colors=always --coverage-clover clover.xml",
-        "upload-coverage": "coveralls -v"
+        "test-coverage": "phpunit --colors=always --coverage-clover clover.xml"
     }
 }
