@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /**
  * @copyright   (c) 2017, Vrok
  * @license     MIT License (http://www.opensource.org/licenses/mit-license.php)
@@ -14,55 +16,46 @@ namespace TexWrapper;
 class Wrapper
 {
     /**
-     * Command to use to execute PDFLaTeX or other engine like LuaLaTeX
+     * Command to use to execute PDFLaTeX or other engine like LuaLaTeX.
      *
      * nonstopmode: shows the errors and does not wait for interaction,
      *     batchmode would show no error/warnings at all
      * output-directory: without it pdflatex would put the files
      *     (*.log, *.aux, *.pdf) into the current working directory
      * file-line-error: show filename:line before error messages
-     * @var string
      */
-    protected $command = 'pdflatex --file-line-error'
+    protected string $command = 'pdflatex --file-line-error'
                 .' --interaction=nonstopmode --output-directory=%dir% %file%';
 
     /**
      * Filename used to store the TeX.
-     *
-     * @var string
      */
-    protected $filename = null;
+    protected string $filename;
 
     /**
      * Flag for the destructor to clean up if true.
-     *
-     * @var bool
      */
-    protected $isTempfile = false;
+    protected bool $isTempfile = false;
 
     /**
-     * List of errors that occured while building the PDF
-     *
-     * @var array
+     * List of errors that occured while building the PDF.
      */
-    protected $errors = [];
+    protected array $errors = [];
 
     /**
      * Console output generated by the latex and post-processing commands.
-     *
-     * @var string
      */
-    protected $log = '';
+    protected string $log = '';
 
     /**
      * Class constructor - stores the given filename.
      *
-     * @param string $filename  (optional) if given the TEX is stored in this file and
-     *     kept, if empty a temporary file is used and deleted on destruction
+     * @param string $filename (optional) if given the TEX is stored in this file and
+     *                         kept, if empty a temporary file is used and deleted on destruction
      */
     public function __construct(string $filename = null)
     {
-        if (! $filename) {
+        if (!$filename) {
             // tempnam requires a dir, we can not use ini_get('upload_tmp_dir')
             // as this would not be set in CLI, sys_get_temp_dir() can probably
             // not be changed to return a vhost specific path,
@@ -90,10 +83,8 @@ class Wrapper
 
     /**
      * Returns the current latex command.
-     *
-     * @return string
      */
-    public function getCommand() : string
+    public function getCommand(): string
     {
         return $this->command;
     }
@@ -103,23 +94,22 @@ class Wrapper
      * within the $PATH environment. Can be used to modify the command, e.g. if
      * you have installed texfot. %dir% and %file% are replaced with the target
      * directory and target filename (including path).
-     *
-     * @param string $cmd
      */
-    public function setCommand(string $cmd)
+    public function setCommand(string $cmd): self
     {
         $this->command = $cmd;
+
+        return $this;
     }
 
     /**
      * Stores the given Tex content in the current file.
      *
-     * @param string $tex
-     * @return boolean      true if the file was written, else false
+     * @return bool true if the file was written, else false
      */
-    public function saveTex(string $tex) : bool
+    public function saveTex(string $tex): bool
     {
-        return file_put_contents($this->filename, $tex) !== false;
+        return false !== file_put_contents($this->filename, $tex);
     }
 
     /**
@@ -133,15 +123,16 @@ class Wrapper
     /**
      * Generates a pdf file from the current LaTeX file.
      *
-     * @return bool     true if the PDF file was written, else false
+     * @return bool true if the PDF file was written, else false
      */
-    public function buildPdf() : bool
+    public function buildPdf(): bool
     {
         $this->errors = [];
 
         if (file_exists($this->filename.'.pdf')) {
-            if (! unlink($this->filename.'.pdf')) {
+            if (!unlink($this->filename.'.pdf')) {
                 $this->errors['pdf'] = 'Old PDF file could not be deleted';
+
                 return false;
             }
         }
@@ -159,9 +150,10 @@ class Wrapper
         exec($command, $output, $return);
 
         // 127 == command not found
-        if ($return == 127) {
+        if (127 == $return) {
             $this->errors['engine'] =
                 'Command not found, engine is not installed or not within the $PATH!';
+
             return false;
         }
 
@@ -169,7 +161,7 @@ class Wrapper
         // latex command returned an exit state > 0 or the PDF wasn't created
         // (e.g. lualatex fails with "fix your writable cache path" but exit
         // code is still 0...)
-        if ($return != 0 || ! file_exists($this->filename.'.pdf')) {
+        if (0 != $return || !file_exists($this->filename.'.pdf')) {
             // return the complete output for debugging
             $this->errors['engine'] = $this->log;
         }
@@ -197,16 +189,13 @@ class Wrapper
      * Allows to execute a custom post-processing command.
      * %dir% and %file% are replaced with the target directory and target
      * filename (including path).
-     *
-     * @param string $command
-     *
-     * @return bool
      */
-    public function postProcess(string $command) : bool
+    public function postProcess(string $command): bool
     {
-        if (! file_exists($this->filename.'.pdf')) {
+        if (!file_exists($this->filename.'.pdf')) {
             $this->errors['postProcessor'] =
                     'PDF file not found, build not started or failed?';
+
             return false;
         }
 
@@ -218,15 +207,17 @@ class Wrapper
         );
 
         exec($postCmd, $outputPost, $returnPost);
-        if ($returnPost != 0) {
+        if (0 != $returnPost) {
             // the post-processing returned an exit state > 0,
             // return the complete output for debugging
             $this->errors['postProcessor'] = implode("\n", $outputPost);
+
             return false;
         }
 
-        if (! file_exists($this->filename.'.pdf')) {
+        if (!file_exists($this->filename.'.pdf')) {
             $this->errors['postProcessor'] = 'post-processing removed the PDF!';
+
             return false;
         }
 
@@ -236,10 +227,8 @@ class Wrapper
     /**
      * Retrieve all messages from the last PDF build (and post-processing) if
      * some fatal error occured.
-     *
-     * @return array
      */
-    public function getErrors() : array
+    public function getErrors(): array
     {
         return $this->errors;
     }
@@ -247,20 +236,16 @@ class Wrapper
     /**
      * Retrieve the console output generated by latex, can contain warnings
      * like missing fonts or smaller tex errors that were ignored by the engine.
-     *
-     * @return string
      */
-    public function getLog() : string
+    public function getLog(): string
     {
         return $this->log;
     }
 
     /**
      * Retrieve the autogenerated or configured file name.
-     *
-     * @return string
      */
-    public function getFilename() :string
+    public function getFilename(): string
     {
         return $this->filename;
     }
@@ -268,11 +253,12 @@ class Wrapper
     /**
      * Returns the name of the generated PDF file.
      *
-     * @return string   the pdf file name or NULL if it was not (yet) generated
+     * @return string the pdf file name or NULL if it was not (yet) generated
      */
-    public function getPdfFile() : ?string
+    public function getPdfFile(): ?string
     {
         $filename = $this->filename.'.pdf';
+
         return file_exists($filename) ? $filename : null;
     }
 }
